techDebt(buRefactorTraining): Move getFileType into whichFile and use in validatePut
