+++ server/routes/establishments/bulkUpload/uploadFiles.js
@@ -2,12 +2,11 @@
 const csv = require('csvtojson');
 
 const config = require('../../../config/config');
-const EstablishmentCsvValidator = require('../../../models/BulkImport/csv/establishments').Establishment;
 const S3 = require('./s3');
 const { buStates } = require('./states');
 const Bucket = S3.Bucket;
 const validatorFactory = require('../../../models/BulkImport/csv/validatorFactory').validatorFactory;
-const { isWorkerFile, isTrainingFile } = require('./whichFile');
+const { getFileType } = require('./whichFile');
 const { validateWorkerHeaders } = require('../bulkUpload/validate/headers/worker');
 const { validateTrainingHeaders } = require('../bulkUpload/validate/headers/training');
 
@@ -179,16 +178,10 @@ const uploadedPut = async (req, res) => {
 
     const allContent = await Promise.all(createModelPromises);
 
-    allContent.forEach((myfile) => {
-      if (EstablishmentCsvValidator.isContent(myfile.data)) {
-        myDownloads.push(createMyFileObject(myfile, 'Establishment'));
-      } else if (isWorkerFile(myfile.data)) {
-        myDownloads.push(createMyFileObject(myfile, 'Worker'));
-      } else if (isTrainingFile(myfile.data)) {
-        myDownloads.push(createMyFileObject(myfile, 'Training'));
-      } else {
-        myDownloads.push(createMyFileObject(myfile, null));
-      }
+    allContent.forEach((file) => {
+      const fileType = getFileType(file.data);
+
+      myDownloads.push(createMyFileObject(file, fileType));
     });
 
     await Promise.all(
